+++ src/main/Helpers/NormalizeInput.php
@@ -4,8 +4,12 @@ namespace Gam\Estafeta\Command\Helpers;
 
 trait NormalizeInput
 {
-    private function normalize(string $content): string
+    private function normalize(?string $content): string
     {
+        if (is_null($content)) {
+            return '';
+        }
+
         $unaccent = Strings::unaccent($content);
         return strtoupper($unaccent);
     }