+++ src/main/Model/Dimension.php
@@ -1,8 +1,13 @@
 <?php
 
+/** @noinspection PhpArrayShapeAttributeCanBeAddedInspection */
+
 namespace Gam\Estafeta\Command\Model;
 
-class Dimension
+use Gam\Estafeta\Command\Validation\ValidatedModel;
+use Nette\Schema\Expect;
+
+class Dimension extends ValidatedModel
 {
     private int $length;
 
@@ -17,9 +22,10 @@ class Dimension
      */
     public function __construct(int $length, int $high, int $width)
     {
-        $this->length = $length;
-        $this->high = $high;
-        $this->width = $width;
+        parent::__construct(compact(['length', 'high', 'width']));
+        $this->length = $this->normalized->length;
+        $this->high = $this->normalized->high;
+        $this->width = $this->normalized->width;
     }
 
     /**
@@ -44,5 +50,14 @@ class Dimension
     public function getWidth(): int
     {
         return $this->width;
+    }
+
+    protected function validationRules(): array
+    {
+        return [
+            'length' => Expect::int()->required()->min(1)->max(150),
+            'high' => Expect::int()->required()->min(1)->max(155),
+            'width' => Expect::int()->required()->min(1)->max(115),
+        ];
     }
 }